Restructure the renderer
